+++ lib/common/common.dart
@@ -1,4 +1,3 @@
-export '../app/login/widgets/login_header.dart';
 export 'app_bar/common_app_bar.dart';
 export 'layouts/product_grid_layout.dart';
 export 'styles/shadow_styles.dart';
