+++ lib/home/widgets/promotion_slider.dart
