+++ lib/onboarding/widgets/onboarding_skip_widget.dart
