+++ lib/shop/widgets/brand_card.dart
